Add 'About' section and update 'Skills' and icons
Added a new 'About' section to the webpage to provide a brief personal introduction. Updated the 'Skills' section to include Git and improved the representation of skills with refactorings. Also, updated SVG icons for JavaScript and Git for a better appearance. Modified navigation; now clicking on the logo redirects to the homepage. Made various minor changes for better alignment and layout.