Add sitemap and corresponding links
A new sitemap has been added to improve SEO and site accessibility. This has been integrated into the Astro configuration file, added as a dependency in package.json, and linked in the Layout component. The robots.txt file has also been updated to allow all user-agents and point to the new sitemap.