Add system and server version info on Info page
Expanded the Info page to display system information and versions of Nginx and Apache. The application will now check the platform (Windows or Unix-based) and display the appropriate system information. It will also attempt to execute commands to obtain versions of Nginx and Apache. If the commands fail, it will display an appropriate error message.